backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
